requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
gevent==23.9.1
werkzeug==3.0.1

# Data Processing & Analysis
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.1

# AWS Services
boto3==1.28.85
botocore==1.31.85

# Riot API & HTTP
requests==2.31.0
ratelimit==2.2.1
requests-cache==1.1.0

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23

# ML & AI
torch>=2.0.0
torchvision>=0.15.0

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
pytz==2023.3
streamlit==1.35.0